+++ packages/components-js/tests/vrt/puppeteer/specs/component-high-contrast-mode.vrt.ts
@@ -41,6 +41,7 @@ const components: Component[] = [
   'spinner',
   'stepper-horizontal',
   'switch',
+  'table',
   'tabs',
   'tabs-bar',
   'tag',